// FL410.Frontend/src/types/IDA.ts
export interface IDA {
    id: string;
    numeroEmenda: number;
    situacao: string;
    produto: string;
    efetividade: string;
    lingua: string;
    revogou: boolean;
    documentos: string;
    sistema: string;
    fabricante: string;
    modelo: string;
    numeroDiretriz: string;

    // Voltar para DA.tsx e implementar a listagem das DAs.

}